Version 2.0
The new release of this fixes some stuff I would label issues: Identifiers displayed as normal text, Statements/Types not fitting with the overall theme, etc. This new release is a bit of a larger update as far as the color palate goes, so be sure to hang onto your original version if the new one is too much of a change. Personally, I think it's a lot nicer and easier on the eyes.